Design: 주차 아이템 타입별 이미지 추가(엄마/아기)
